ToChange project return data type

+++ app/src/main/java/com/gabrielMJr/twaire/physic/units/VolumeIF.java
@@ -1,12 +1,14 @@
 package com.gabrielMJr.twaire.physic.units;
 
+import java.math.BigDecimal;
+
 public interface VolumeIF
 {
-	public abstract Double toCubeKilometer(double value, String unit);
-	public abstract Double toCubeHectometer(double value, String unit);
-	public abstract Double toCubeDecameter(double value, String unit);
-	public abstract Double toCubeMeter(double value, String unit);
-	public abstract Double toCubeDecimeter(double value, String unit);
-	public abstract Double toCubeCentimeter(double value, String unit);
-	public abstract Double toCubeMillimeter(double value, String unit);
+	public abstract BigDecimal toCubeKilometer(double value, String unit);
+	public abstract BigDecimal toCubeHectometer(double value, String unit);
+	public abstract BigDecimal toCubeDecameter(double value, String unit);
+	public abstract BigDecimal toCubeMeter(double value, String unit);
+	public abstract BigDecimal toCubeDecimeter(double value, String unit);
+	public abstract BigDecimal toCubeCentimeter(double value, String unit);
+	public abstract BigDecimal toCubeMillimeter(double value, String unit);
 }
